restructure to hide templates

+++ src/create_run.sh
@@ -9,8 +9,12 @@ function create_run() {
         echo "Error: Argument is empty!"
         return 1
     fi
+    if [ -f $1 ]; then
+        echo "Run directory $PWD/$1 already exists!"
+        return 1
+    fi
 
-    ln -s $MAPSEQ_REPO_PATH/LRAXXX_template $1
+    ln -s $MAPSEQ_REPO_PATH/src/LRAXXX_template $1
     echo "Lane,Sample,Index" > $1/data/$1.ATAC.sampleManifest.csv
     echo "Lane,Sample,Index" > $1/data/$1.RNA.sampleManifest.csv
     cp $MAPSEQ_REPO_PATH/src/references/HTOB_feature_ref.csv $1/pipeline/ATAC.ASAP/HTOB_feature_ref.csv
@@ -31,6 +35,10 @@ function create_aggr_run() {
         echo "Error: Argument is empty!"
         return 1
     fi
+    if [ -f $1 ]; then
+        echo "Run directory $PWD/$1 already exists!"
+        return 1
+    fi
 
-    ln -s $MAPSEQ_REPO_PATH/LRA_all_template $1
+    ln -s $MAPSEQ_REPO_PATH/src/LRA_all_template $1
 }
